backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
snapgene-reader>=0.1.20
biopython>=1.82
ViennaRNA>=2.6.0
scipy>=1.11.0
optuna>=3.4.0
matplotlib>=3.8.0
pandas>=2.1.0
requests>=2.31.0
